rootfs: preserve symlinks for some conffile diversions
The thing is that some packages may destroy symlinks chain,
for example:

 * libnss-mdns uses inline Perl editing in postinst, which replaces
   symlink with a regular file;
 * kmod may copy resulting conffile to its destination.

So we should install wb-configs after these packages to preserve
symlinks in conffiles diversion, or wb-configs postinst script will
generate warnings about symlinks (and it also breaks wb-configs
behavior).

This may be fixed in config-package-dev helpers, but for now we
don't want to mess up with it. Maybe there will be a fix for
wb-configs itself.

Signed-off-by: Nikita webconn Maslov <n.maslov@contactless.ru>