CMake support (#67)
* CMake support

We provide our own versions of libcomerr2 and e2fslibs for armel and armhf, but not for amd64. So we can't use them in multiarch mode during crosscompilation with amd64 system. CMake is dependend of these libraries. In order to support CMake distr supplied libraries are installed in devenv.

* Some dev libraries are installed in devenv from upstream

We don't provide amd64 versions of some modified libraries. To avoid errors during cross build upstream versions are installed.