devenv: force update debian-archive-keyring in Docker
Current Debian buster (and bullseye) repository is signed with a new key
which is not available in debian-archive-keyring from stretch.

So, in order to build rootfs for the docker image we need this new
debian-archive-keyring package. It is already backported from bullseye
to our dev-tools repository, so it's enough to install/update it manually
on docker image build.